+++ Core/AppURLs.swift
@@ -82,7 +82,7 @@ public extension URL {
         return getParameter(named: Param.search)
     }
 
-    enum Param {
+    fileprivate enum Param {
 
         static let search = "q"
         static let source = "t"
@@ -99,7 +99,7 @@ public extension URL {
 
     }
 
-    enum ParamValue {
+    fileprivate enum ParamValue {
 
         static let source = "ddg_ios"
         static let appUsage = "app_use"
@@ -123,6 +123,137 @@ public extension URL {
     func removingInternalSearchParameters() -> URL {
         guard isDuckDuckGoSearch else { return self }
         return removingParameters(named: [Param.atb, Param.source, Param.searchHeader])
+    }
+
+    // MARK: - StatisticsDependentURLs
+
+    private static let defaultStatisticsDependentURLFactory = StatisticsDependentURLFactory()
+
+    static func makeSearchURL(text: String) -> URL? { defaultStatisticsDependentURLFactory.makeSearchURL(text: text) }
+
+    static func makeSearchURL(query: String, queryContext: URL? = nil) -> URL? {
+        defaultStatisticsDependentURLFactory.makeSearchURL(query: query, queryContext: queryContext)
+    }
+
+    func applyingStatsParams() -> URL { URL.defaultStatisticsDependentURLFactory.applyingStatsParams(to: self) }
+
+    static var searchAtb: URL? { defaultStatisticsDependentURLFactory.searchAtb }
+
+    static var appAtb: URL? { defaultStatisticsDependentURLFactory.appAtb }
+
+    var hasCorrectMobileStatsParams: Bool { URL.defaultStatisticsDependentURLFactory.hasCorrectMobileStatsParams(url: self) }
+
+    static func makePixelURL(pixelName: String, formFactor: String? = nil, includeATB: Bool = true) -> URL {
+        defaultStatisticsDependentURLFactory.makePixelURL(pixelName: pixelName, formFactor: formFactor, includeATB: includeATB)
+    }
+
+}
+
+public final class StatisticsDependentURLFactory {
+
+    private let statisticsStore: StatisticsStore
+
+    init(statisticsStore: StatisticsStore = StatisticsUserDefaults()) {
+        self.statisticsStore = statisticsStore
+    }
+
+    // MARK: Search
+
+    func makeSearchURL(text: String) -> URL? {
+        makeSearchURL(text: text, additionalParameters: [])
+    }
+
+    func makeSearchURL(query: String, queryContext: URL? = nil) -> URL? {
+        if let url = URL.webUrl(from: query) {
+            return url
+        }
+
+        var parameters = [String: String]()
+        if let queryContext = queryContext,
+           queryContext.isDuckDuckGoSearch,
+           queryContext.getParameter(named: URL.Param.verticalMaps) == nil,
+           let vertical = queryContext.getParameter(named: URL.Param.vertical),
+           URL.ParamValue.majorVerticals.contains(vertical) {
+
+            parameters[URL.Param.verticalRewrite] = vertical
+        }
+
+        return makeSearchURL(text: query, additionalParameters: parameters)
+    }
+
+    /**
+     Generates a search url with the source (t) https://duck.co/help/privacy/t
+     and cohort (atb) https://duck.co/help/privacy/atb
+     */
+    private func makeSearchURL<C: Collection>(text: String, additionalParameters: C) -> URL
+    where C.Element == (key: String, value: String) {
+        let searchURL = URL.ddg
+            .appendingParameter(name: URL.Param.search, value: text)
+            .appendingParameters(additionalParameters)
+        return applyingStatsParams(to: searchURL)
+    }
+
+    func applyingStatsParams(to url: URL) -> URL {
+        var searchURL = url.removingParameters(named: [URL.Param.source, URL.Param.atb])
+            .appendingParameter(name: URL.Param.source,
+                                value: URL.ParamValue.source)
+
+        if let atbWithVariant = statisticsStore.atbWithVariant {
+            searchURL = searchURL.appendingParameter(name: URL.Param.atb, value: atbWithVariant)
+        }
+        return searchURL
+    }
+
+    // MARK: ATB
+
+    var searchAtb: URL? {
+        guard let atbWithVariant = statisticsStore.atbWithVariant, let setAtb = statisticsStore.searchRetentionAtb else {
+            return nil
+        }
+        return URL.atb.appendingParameters([
+            URL.Param.atb: atbWithVariant,
+            URL.Param.setAtb: setAtb,
+            URL.Param.email: EmailManager().isSignedIn ? URL.ParamValue.emailEnabled : URL.ParamValue.emailDisabled
+        ])
+    }
+
+    var appAtb: URL? {
+        guard let atbWithVariant = statisticsStore.atbWithVariant, let setAtb = statisticsStore.appRetentionAtb else {
+            return nil
+        }
+        return URL.atb.appendingParameters([
+            URL.Param.activityType: URL.ParamValue.appUsage,
+            URL.Param.atb: atbWithVariant,
+            URL.Param.setAtb: setAtb,
+            URL.Param.email: EmailManager().isSignedIn ? URL.ParamValue.emailEnabled : URL.ParamValue.emailDisabled
+        ])
+    }
+
+    func hasCorrectMobileStatsParams(url: URL) -> Bool {
+        guard let source = url.getParameter(named: URL.Param.source),
+              source == URL.ParamValue.source
+        else { return false }
+        if let atbWithVariant = statisticsStore.atbWithVariant {
+            return atbWithVariant == url.getParameter(named: URL.Param.atb)
+        }
+        return true
+    }
+
+    // MARK: Pixel
+
+    private static let pixelBase: String = ProcessInfo.processInfo.environment["PIXEL_BASE_URL", default: "https://improving.duckduckgo.com"]
+    func makePixelURL(pixelName: String, formFactor: String? = nil, includeATB: Bool = true) -> URL {
+        var urlString = "\(Self.pixelBase)/t/\(pixelName)"
+        if let formFactor = formFactor {
+            urlString.append("_ios_\(formFactor)")
+        }
+        var url = URL(string: urlString)!
+
+        if includeATB {
+            url = url.appendingParameter(name: URL.Param.atb, value: statisticsStore.atbWithVariant ?? "")
+        }
+
+        return url
     }
 
 }
